Refactor incidents
