+++ Client/src/Pages/Incidents/Components/OptionsHeader/skeleton.jsx
@@ -1,0 +1,11 @@
+import { Stack, Skeleton } from "@mui/material";
+
+const SkeletonLayout = () => {
+	return (
+		<Stack>
+			<Skeleton height={40} />
+		</Stack>
+	);
+};
+
+export default SkeletonLayout;
