--- Client/src/Pages/Incidents/IncidentTable/Skeleton/Skeleton.jsx
@@ -1,21 +1,0 @@
-import { Skeleton /* , Stack */ } from "@mui/material";
-const IncidentSkeleton = () => {
-	return (
-		<>
-			<Skeleton
-				animation={"wave"}
-				variant="rounded"
-				width="100%"
-				height={300}
-			/>
-			<Skeleton
-				animation={"wave"}
-				variant="rounded"
-				width="100%"
-				height={100}
-			/>
-		</>
-	);
-};
-
-export { IncidentSkeleton };
