--- Client/src/Pages/Uptime/Monitors/Components/UptimeDataTable/skeleton.jsx
@@ -1,14 +1,0 @@
-import { Skeleton } from "@mui/material";
-
-const UptimeDataTableSkeleton = () => {
-	return (
-		<Skeleton
-			variant="rounded"
-			width="100%"
-			height="100%"
-			flex={1}
-		/>
-	);
-};
-
-export default UptimeDataTableSkeleton;
